requirements.txt
django 
djangorestframework 
django-environ
psycopg==3.1.12
djangorestframework-simplejwt 
django-cors-headers

